+++ src/Information/CorrectedMutualInformation.jl
@@ -57,7 +57,7 @@ function buslje09(aln::AbstractMatrix{Residue};
         zmi = copy(mi)
         residuematrix = getresidues(aln)
         for ns in 1:samples
-            shuffle!(residuematrix,1,fixedgaps)
+            shuffle_msa!(residuematrix, dims=1, fixedgaps=fixedgaps)
             rand_mi[ns] = getarray(_buslje09(aln, alphabet, clusters, lambda, apc))
         end
         PairwiseListMatrices.zscore!(rand_mi, getarray(zmi))
@@ -132,7 +132,7 @@ function BLMI(aln::AbstractMatrix{Residue};
         zmi = copy(mi)
         residuematrix = getresidues(aln)
         for ns in 1:samples
-            shuffle!(residuematrix,1,fixedgaps)
+            shuffle_msa!(residuematrix, dims=1, fixedgaps=fixedgaps)
             rand_mi[ns] = getarray(_BLMI(aln, clusters, numbercl, beta, apc, lambda))
         end
         PairwiseListMatrices.zscore(rand_mi, getarray(zmi))